+++ CHANGES.rst
@@ -4,8 +4,8 @@ Changelog
 2.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Handle also limit in querystringsearch patch.
+  [cekk]
 
 2.0.1 (2021-05-05)
 ------------------